changed the imageuploader and image preview modal to a component
